requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.2
cycler==0.12.1
einops==0.8.1
filelock==3.19.1
fire==0.7.1
fonttools==4.60.1
frozenlist==1.7.0
fsspec==2025.9.0
gitdb==4.0.12
gitpython==3.1.45
idna==3.10
jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lightning-utilities==0.15.2
markupsafe==3.0.3
matplotlib==3.10.6
monai==1.5.1
mpmath==1.3.0
multidict==6.6.4
networkx==3.4.2
nodeenv==1.9.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.4.0
propcache==0.3.2
protobuf==6.32.1
pydantic==2.11.9
pydantic-core==2.33.2
pyparsing==3.2.5
pyright==1.1.406
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.5
pytz==2025.2
pyyaml==6.0.3
requests==2.32.5
ruff==0.13.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
sentry-sdk==2.39.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sympy==1.14.0
termcolor==3.1.0
threadpoolctl==3.6.0
torch==2.8.0
torchmetrics==1.8.2
tqdm==4.67.1
triton==3.4.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
-e file:///home/mgazda/Projects/UMBRA
urllib3==2.5.0
wandb==0.22.1
yarl==1.20.1
